+++ test/credo/check/readability/pipe_into_anonymous_functions_test.exs
@@ -1,0 +1,64 @@
+defmodule Credo.Check.Readability.PipeIntoAnonymousFunctionsTest do
+  use Credo.Test.Case
+
+  @described_check Credo.Check.Readability.PipeIntoAnonymousFunctions
+
+  #
+  # cases NOT raising issues
+  #
+
+  test "it should NOT report expected code" do
+    """
+    defmodule CredoSampleModule do
+      use ExUnit.Case
+
+      def some_fun do
+        some_val
+        |> do_something
+        |> do_something_else
+      end
+    end
+    """
+    |> to_source_file
+    |> run_check(@described_check)
+    |> refute_issues()
+  end
+
+  #
+  # cases raising issues
+  #
+
+  test "it should report a violation" do
+    """
+    defmodule CredoSampleModule do
+      use ExUnit.Case
+
+      def some_fun do
+        some_val
+        |> (fn x -> x * 2 end).()
+        |> do_something
+      end
+    end
+    """
+    |> to_source_file
+    |> run_check(@described_check)
+    |> assert_issue()
+  end
+
+  test "it should report a violation for multiple violations" do
+    """
+    defmodule CredoSampleModule do
+      use ExUnit.Case
+
+      def some_fun do
+        some_val
+        |> (fn x -> x * 2 end).()
+        |> (fn x -> x * 2 end).()
+      end
+    end
+    """
+    |> to_source_file
+    |> run_check(@described_check)
+    |> assert_issues()
+  end
+end